+++ react/components/SearchResult.js
@@ -16,13 +16,13 @@ class SearchResult extends Component {
 
   static defaultProps = {
     hiddenFacets: {
-      layoutMode1: LAYOUT_MODE[0].value,
-      layoutMode2: LAYOUT_MODE[1].value,
+      layoutMode1: LAYOUT_MODE[1].value,
+      layoutMode2: LAYOUT_MODE[0].value,
     },
   }
 
   state = {
-    galleryLayoutMode: this.props.hiddenFacets.layoutMode1 || LAYOUT_MODE[0].value,
+    mobileLayoutMode: this.props.hiddenFacets.layoutMode1 || LAYOUT_MODE[0].value,
     showLoadingAsOverlay: false,
     // The definitions bellow are required because
     // on SSR the getDerivedStateFromProps isn't called
@@ -39,15 +39,15 @@ class SearchResult extends Component {
     hiddenFacets: this.props.hiddenFacets,
   }
 
-  handleLayoutChange = e => {
+  handleMobileLayoutChange = e => {
     e.preventDefault()
 
     const defaultModes = [this.props.hiddenFacets.layoutMode1, this.props.hiddenFacets.layoutMode2]
-    const modeIndex = (defaultModes.indexOf(this.state.galleryLayoutMode) + 1) % 2
+    const modeIndex = (defaultModes.indexOf(this.state.mobileLayoutMode) + 1) % 2
     const currentMode = defaultModes[modeIndex]
 
     this.setState({
-      galleryLayoutMode: currentMode,
+      mobileLayoutMode: currentMode,
     })
   }
 
@@ -121,7 +121,7 @@ class SearchResult extends Component {
       runtime: { hints: { mobile } },
     } = this.props
     const {
-      galleryLayoutMode,
+      mobileLayoutMode,
       recordsFiltered,
       products,
       brands,
@@ -188,7 +188,7 @@ class SearchResult extends Component {
                 products={products}
                 summary={summary}
                 className="bn"
-                layoutMode={galleryLayoutMode}
+                mobileLayoutMode={mobileLayoutMode}
               />
             )}
             {children}
@@ -202,8 +202,8 @@ class SearchResult extends Component {
           {mobile && <div className={`${searchResult.switch} flex justify-center items-center`}>
             <div className="dn-ns db-s">
               <LayoutModeSwitcher
-                activeMode={galleryLayoutMode}
-                onChange={this.handleLayoutChange}
+                activeMode={mobileLayoutMode}
+                onChange={this.handleMobileLayoutChange}
               />
             </div>
           </div>}