+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix misuse of `layoutMode` prop in desktop view.
+
+### Changed
+- Rename `layoutMode` to `mobileLayoutMode` to avoid using it in desktop view.
 
 ## [3.11.6] - 2019-02-12
 ### Fixed